requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.21.3
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.0.6
